refactor(chat): 抽出 splitThink 模块并修嵌套标签重复
响应 PR #536 的 CodeRabbit review：

1. (CHANGES_REQUESTED) splitThink 在循环里跳过 idx < cursor 的开标签，避免嵌套/重叠  标签导致内容重复(inner 出现两次)。
2. (Nitpick) 把纯解析逻辑 Segment/splitThink 抽到独立模块 splitThink.ts，MarkdownContent.tsx 与测试改从该模块 import——这样 splitThink 单测不再加载 ReactMarkdown/remark-gfm/lucide-react。
3. 新增嵌套标签回归用例 splitThink.test.ts，验证内容不重复 + 尾部文本保留。
